Multiple changes. Click for details
Removed follow friend method because i thought it was going to be on home page? Changed friend list to be a drop down of friends. If a friend is selected, you can hit the "Go" button to be redirected to their page. As of you cannot view another person's page because friendsPage.jsp is unfinished